requirements.txt
certifi==2018.8.13
chardet==3.0.4
diff-match-patch==20121119
Django>=2.1.6
django-js-asset==1.2.1
django-mptt==0.9.1
et-xmlfile==1.0.1
idna==2.7
isodate==0.6.0
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.5
Pillow==5.4.1
psycopg2-binary==2.7.5
python-dateutil==2.8.0
pytz==2018.5
PyYAML==5.1
requests==2.21.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.25.1
vobject==0.9.6.1
xlrd==1.1.0
xlwt==1.3.0
